+++ src/com/intellij/idea/plugin/hybris/system/type/psi/reference/result/OrderingAttributeResolveResult.kt
@@ -1,0 +1,31 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.intellij.idea.plugin.hybris.system.type.psi.reference.result
+
+import com.intellij.idea.plugin.hybris.psi.reference.TSReferenceBase
+import com.intellij.idea.plugin.hybris.system.type.meta.model.TSMetaRelation
+import com.intellij.idea.plugin.hybris.system.type.model.RelationElement
+
+class OrderingAttributeResolveResult(
+    val meta: TSMetaRelation.TSMetaOrderingAttribute
+) : TSReferenceBase.TSResolveResult {
+    private val myDom: RelationElement? = meta.retrieveDom()
+    override fun getElement() = myDom?.qualifier?.xmlAttributeValue
+    override fun isValidResult() = (myDom?.isValid ?: false) && element != null
+}